Fix out of bounds access when generating temporary files in CUDA RTC interface
The CUDA RTC interface in tc/core/cuda/cuda_rtc.cc generates temporary
files through mkstemp. The filename pattern passed to mkstemp is the
address of a sequence of characters stored in a vector. This vector is
initialized with the characters of a string containing the pattern:

  std::string pat("/tmp/cudaXXXXXX");
  std::vector<char> ifn(pat.begin(), pat.end());
  TC_CHECK_GE(mkstemp(ifn.data()), 0); // string.c_str is const char*
  std::string inputFileName(ifn.begin(), ifn.end());

Since the vector is initialized with a range of iterators on the
string with the pattern, it does not contain a terminating NUL
character. However, mkstemp requires the sequence of characters for
the pattern to be zero-terminated. Hence, simply passing the data
pointer of the vector to mkstemp potentially results in out of bounds
accesses.

This patch fixes the problem by directly passing a zero-terminated
sequence of characters allocated on the stack to mkstemp and by
generating the string with the filename directly from the modified
sequence:

  char pat[] = "/tmp/cudaXXXXXX";
  TC_CHECK_GE(mkstemp(pat), 0);
  std::string inputFileName(pat);