Close temporary files generated by mkstemp in CUDA RTC interface
The function mkstemp both creates and opens a temporary file and
returns a file descriptor to the caller. Currently, the CUDA RTC
interface deletes temporary files, but does not close the file
descriptors. Temporary files are thus marked for deletion, but remain
open. This may result in too many open files and cause the process to
be killed by the operating system.

This patch properly closes all valid file descriptors returnd by
mkstemp.